chore: move directory of components only used in contest
