+++ apps/frontend/app/admin/contest/[id]/edit/page.tsx
@@ -1,9 +1,7 @@
 'use client'
 
-import ContestProblemListLabel from '@/app/admin/_components/ContestProblemListLabel'
 import DescriptionForm from '@/app/admin/_components/DescriptionForm'
 import FormSection from '@/app/admin/_components/FormSection'
-import ImportProblemButton from '@/app/admin/_components/ImportProblemButton'
 import SwitchField from '@/app/admin/_components/SwitchField'
 import TitleForm from '@/app/admin/_components/TitleForm'
 import { DataTableAdmin } from '@/components/DataTableAdmin'
@@ -30,6 +28,8 @@ import { FormProvider, useForm } from 'react-hook-form'
 import { FaAngleLeft } from 'react-icons/fa6'
 import { IoMdCheckmarkCircleOutline } from 'react-icons/io'
 import { toast } from 'sonner'
+import ContestProblemListLabel from '../../_components/ContestProblemListLabel'
+import ImportProblemButton from '../../_components/ImportProblemButton'
 import TimeForm from '../../_components/TimeForm'
 import { type ContestProblem, editSchema } from '../../utils'
 import { columns } from '../_components/Columns'
